printwidth 100

+++ src/api/admin/requests.ts
@@ -1,9 +1,4 @@
-import {
-  useMutation,
-  useQuery,
-  useQueryClient,
-  UseQueryOptions
-} from "react-query"
+import { useMutation, useQuery, useQueryClient, UseQueryOptions } from "react-query"
 import axios, { AxiosRequestConfig } from "axios"
 import { isEmpty, isNumber, keyBy, pick, property } from "lodash-es"
 
@@ -12,11 +7,7 @@ import { ParentPost } from "../Post"
 import { ServiceMessage } from "../ServiceMessage"
 import { challengeIdParam } from "../requests"
 
-import {
-  AdminChallengeDict,
-  AdminChallengePayload,
-  ChallengePreview
-} from "./Challenge"
+import { AdminChallengeDict, AdminChallengePayload, ChallengePreview } from "./Challenge"
 import { AdminServiceMessagePayload } from "./ServiceMessage"
 
 const getChallenges = async () =>
@@ -24,11 +15,10 @@ const getChallenges = async () =>
 export const useChallenges = <TSelected = AdminChallengeDict>(
   options?: UseQueryOptions<AdminChallengeDict, QueryError, TSelected>
 ) =>
-  useQuery<AdminChallengeDict, QueryError, TSelected>(
-    ["admin", "challenges"],
-    getChallenges,
-    { ...options, staleTime: 600_000 }
-  )
+  useQuery<AdminChallengeDict, QueryError, TSelected>(["admin", "challenges"], getChallenges, {
+    ...options,
+    staleTime: 600_000
+  })
 export const useChallenge = (door: number | null | undefined) =>
   useQuery<AdminChallengeDict, QueryError, AdminChallengeDict[number]>(
     ["admin", "challenges"],
@@ -37,25 +27,17 @@ export const useChallenge = (door: number | null | undefined) =>
   )
 
 const getPosts = (door: number) =>
-  axios
-    .get(`/admin/challenges/${challengeIdParam(door)}/posts`)
-    .then(({ data }) => data)
+  axios.get(`/admin/challenges/${challengeIdParam(door)}/posts`).then(({ data }) => data)
 export const usePosts = (door: number) =>
-  useQuery<ParentPost[], QueryError>(
-    ["admin", "posts", door],
-    () => getPosts(door),
-    { staleTime: 300_000 }
-  )
+  useQuery<ParentPost[], QueryError>(["admin", "posts", door], () => getPosts(door), {
+    staleTime: 300_000
+  })
 
-export const getChallengePreview = async (
-  challenge: AdminChallengePayload | undefined
-) => {
+export const getChallengePreview = async (challenge: AdminChallengePayload | undefined) => {
   challenge = pick(challenge, ["markdown_content", "files"])
   if (isEmpty(challenge)) return
 
-  return await axios
-    .post("/admin/challenge_markdown", { challenge })
-    .then(({ data }) => data)
+  return await axios.post("/admin/challenge_markdown", { challenge }).then(({ data }) => data)
 }
 export const useChallengePreview = (
   challenge: AdminChallengePayload | undefined,
@@ -109,8 +91,7 @@ export const useDeleteChallenge = () => {
   return useMutation<unknown, QueryError, DeleteChallengeParameters>(
     ["admin", "challenges", "destroy"],
     async ({ door }) =>
-      isNumber(door) &&
-      axios.delete(`/admin/challenges/${challengeIdParam(door)}`),
+      isNumber(door) && axios.delete(`/admin/challenges/${challengeIdParam(door)}`),
     {
       onSuccess: () => {
         queryClient.invalidateQueries("challenges")
@@ -126,11 +107,7 @@ export type CreateServiceMessageParameters = {
 export const useCreateServiceMessage = () => {
   const queryClient = useQueryClient()
 
-  return useMutation<
-    ServiceMessage,
-    QueryError,
-    CreateServiceMessageParameters
-  >(
+  return useMutation<ServiceMessage, QueryError, CreateServiceMessageParameters>(
     ["admin", "serviceMessages", "create"],
     (data) => axios.post("/admin/service_messages", data),
     {
@@ -148,11 +125,7 @@ export type UpdateServiceMessageParameters = {
 export const useUpdateServiceMessage = () => {
   const queryClient = useQueryClient()
 
-  return useMutation<
-    ServiceMessage,
-    QueryError,
-    UpdateServiceMessageParameters
-  >(
+  return useMutation<ServiceMessage, QueryError, UpdateServiceMessageParameters>(
     ["admin", "serviceMessages", "update"],
     ({ uuid, ...data }) => axios.patch(`/admin/service_messages/${uuid}`, data),
     {
@@ -203,9 +176,7 @@ export const useCreateBlob = () =>
     unknown,
     { blob: CreateBlobPayload; config?: AxiosRequestConfig }
   >(["admin", "activeStorage", "createBlob"], ({ blob, config }) =>
-    axios
-      .post("/rails/active_storage/direct_uploads", { blob }, config)
-      .then(({ data }) => data)
+    axios.post("/rails/active_storage/direct_uploads", { blob }, config).then(({ data }) => data)
   )
 
 type UploadFilePayload = {
@@ -214,11 +185,7 @@ type UploadFilePayload = {
 }
 
 export const useUploadFile = () =>
-  useMutation<
-    never,
-    unknown,
-    { upload: UploadFilePayload; config?: AxiosRequestConfig }
-  >(
+  useMutation<never, unknown, { upload: UploadFilePayload; config?: AxiosRequestConfig }>(
     ["admin", "activeStorage", "uploadFile"],
     ({ upload: { file, directUpload }, config }) =>
       axios.put(directUpload.url, file, {
